Do what it takes to solve Day 8 part 2.
Algorithm still doesn't complete, but the answer is still printed because everything we need was quickly entered into the memoization table.

Assumes that all iterations are prime numbers of full trips though the instructions, which are also of prime count. Also assumes that the number of iterations between the start node and the end node is the same as between the end node and itself.